hw2_flask/requirements.txt
psycopg2-binary==2.9.5
Flask==2.2.2
python-dotenv==0.21.0
